// CourierAmexClient/src/app/features/package/models/PreStudy-PreAlert.model.ts
export interface ManifestPreAlert {
    gateway: string;
    account: string;
    aipe: string;
    crtrack: number;
    courierCode: string;
    description: string;
    shipper: string;
    consignee: string;
    company: string;
    identification: string;
    typeId: string;
    value: number;
    weight: number;
    invoice: string;
    xtn: string;
    address: string;
    city: string;
    country: string;
    telephone: string;
    exonerate: string;
}